issue-16 create ApiClient class and move PaymentIntent creation and API request logic there
